+++ xmla/CHANGES.md
@@ -6,7 +6,7 @@ CHANGES
 * ``as_user`` and ``spn`` are no longer ignored in the kerberos authentication
 * implemented the procedural interface from olap.interfaces
 * fixed problem when no sliceraxis info is returned
-* parmeter ``property`` of getSlice now spells ``properties``
+* parameter ``property`` of getSlice now spells ``properties``
 
 0.4
 ----